requirements.txt
kafka-python==2.2.15
pyspark==3.5.1
py4j==0.10.9.7
pandas==2.0.3
numpy==1.24.3 